# ARROW-13987: [C++] Support nested field refs