--- a/cpp/src/arrow/array/array_nested.h
+++ b/cpp/src/arrow/array/array_nested.h
@@ -370,12 +370,13 @@ class ARROW_EXPORT StructArray : public Array {
   /// \param[in] pool The pool to allocate null bitmaps from, if necessary
   Result<ArrayVector> Flatten(MemoryPool* pool = default_memory_pool()) const;
 
-  /// \brief Get one of the child arrays, adjusting the null bitmap if necessary.
+  /// \brief Get one of the child arrays, combining its null bitmap
+  /// with the parent struct array's bitmap.
   ///
   /// \param[in] index Which child array to get
   /// \param[in] pool The pool to allocate null bitmaps from, if necessary
-  Result<std::shared_ptr<Array>> Flatten(int index,
-                                         MemoryPool* pool = default_memory_pool()) const;
+  Result<std::shared_ptr<Array>> GetFlattenedField(
+      int index, MemoryPool* pool = default_memory_pool()) const;
 
  private:
   // For caching boxed child data