--- a/cpp/src/arrow/compute/exec/expression.cc
+++ b/cpp/src/arrow/compute/exec/expression.cc
@@ -398,9 +398,7 @@ Result<Expression> BindImpl(Expression expr, const TypeOrSchema& in,
 
     auto bound = *expr.parameter();
     bound.indices.resize(path.indices().size());
-    for (size_t i = 0; i < path.indices().size(); ++i) {
-      bound.indices[i] = path.indices()[i];
-    }
+    std::copy(path.indices().begin(), path.indices().end(), bound.indices.begin());
     ARROW_ASSIGN_OR_RAISE(auto field, path.Get(in));
     bound.descr.type = field->type();
     bound.descr.shape = shape;
@@ -530,8 +528,8 @@ Result<Datum> ExecuteScalarExpression(const Expression& expr, const ExecBatch& i
         field = struct_scalar.value[index];
       } else if (field.is_array()) {
         const auto& struct_array = field.array_as<StructArray>();
-        ARROW_ASSIGN_OR_RAISE(field,
-                              struct_array->Flatten(index, exec_context->memory_pool()));
+        ARROW_ASSIGN_OR_RAISE(
+            field, struct_array->GetFlattenedField(index, exec_context->memory_pool()));
       } else {
         return Status::NotImplemented("Nested field reference into a ", field.ToString());
       }