@lidavidm 2021-10-19
```
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
Before:
BindAndEvaluate/simple_array         287 ns          287 ns      2384362
BindAndEvaluate/simple_scalar        283 ns          283 ns      2459757
After:
BindAndEvaluate/simple_array         285 ns          285 ns      2470618
BindAndEvaluate/simple_scalar        272 ns          272 ns      2574147
BindAndEvaluate/nested_array        1022 ns         1021 ns       685364
BindAndEvaluate/nested_scalar        632 ns          631 ns      1128511

```