@lidavidm 2021-10-19
Note that `nested_array` is particularly bad (extracting from a StructArray) because we may have to allocate a new validity bitmap for the child, and we may have to do so multiple times (once per nesting level).