@lidavidm 2021-11-05
I was (slowly) working on union support but I'll split that out and get the logic here into a kernel when I get a chance.