@lidavidm 2021-11-10
I filed ARROW-14658.