@jorisvandenbossche 2021-11-10
We have this line in the scanner code in C++:

https://github.com/apache/arrow/blob/575a4376bb02ba79327eaa61b024f4db9b4a7248/cpp/src/arrow/dataset/scanner_internal.h#L214

which is what I ran into when naively trying to call this in Python (after adding a python binding for creating a nested FieldRef).