@lidavidm 2021-11-10
Ah, sorry, I should've remembered to test that. I'll take a look at this today.